路费取整

+++ src/main/java/com/bestfeng/dydj/controller/TravelController.java
@@ -28,7 +28,7 @@ public class TravelController {
 
     @PostMapping("/fare")
     @ApiOperation("出行价格")
-    public CommonResult<Double> travelFate(@RequestBody TravelRequest request) {
+    public CommonResult<Long> travelFate(@RequestBody TravelRequest request) {
         return CommonResult.success(travelServiceSupport.travelFare(TravelTypeEnums.ofValue(request.getType()), request.getDistance()));
     }
 }
